+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # 2.22.3 (unreleased)
 
 * API:
+    * Add `GET /admin/v2/users-csv/` (\#3276).
     * Fix handling of edge case upon updating user's `project_dirs` (\#3270).
 * Internal:
     * Improve type hints for endpoint functions (\#3280).